基本实现所有功能

+++ sypets-frontend/src/views/admin/User.vue
@@ -3,22 +3,21 @@
      <el-card>
        <el-row :gutter="20">
          <el-col :span="6">
-           <el-input placeholder="输入用户ID" v-model="query" clearable @clear="getUserList">
-             <el-button slot="append" icon="el-icon-search" @click="findOne"></el-button>
+           <el-input placeholder="输入搜索内容" v-model="queryInfo.query" clearable @clear="getUserList">
+             <el-button slot="append" icon="el-icon-search" @click="getUserList"></el-button>
            </el-input>
          </el-col>
        </el-row>
      </el-card>
-     <div v-if="userlist.length===0" class="h-50 text-center text-muted fs-4 fw-bold ">
-       空空如也
-     </div>
 <!--     表格-->
     <table class="table table-striped table-bordered">
       <thead>
         <tr>
+          <th class="text-center">Id</th>
           <th class="text-center">用户名</th>
           <th class="text-center">性别</th>
           <th class="text-center">邮箱地址</th>
+          <th class="text-center">个人简介</th>
           <th class="text-center">头像</th>
           <th class="text-center">状态</th>
           <th class="text-center">操作</th>
@@ -26,28 +25,34 @@
       </thead>
       <tbody>
         <tr v-for="user in userlist" :key="user.id">
-          <td class="tableUint lh-lg text-center" style="cursor: pointer" @click="link(user.username)">{{user.username}}</td>
-          <td class="tableUint lh-lg text-truncate text-center">{{user.sex===0?'未知':(user.sex===1?'男':'女')}}</td>
-          <td class="lh-lg text-center" style="width: 200px">{{user.email}}</td>
-          <td style="width: 200px" class="text-center">
-            <a :href="user.avatar" target="_blank">
-              <img :src="user.avatar" alt="..." style="width: 100px;height: 100px" class="object-fit-cover rounded-1">
-            </a>
+          <td class="tableUint">{{user.id}}</td>
+          <td class="tableUint" style="cursor: pointer" @click="link(user.username)">{{user.username}}</td>
+          <td class="tableUint">{{user.sex===0?'未知':(user.sex===1?'男':'女')}}</td>
+          <td class="tableUint">{{user.email}}</td>
+          <td class="tableUint">{{user.privateInfo}}</td>
+          <td class="tableUint">
+            <a :href="user.avatar" target="_blank">{{user.avatar}}</a>
           </td>
           <td style="width: 100px" :class="user.status===1?'text-success':'text-danger'">{{user.status===1?'正常':'异常'}}</td>
           <td class="text-center">
-              <el-button type="warning" class="my-2" @click="userStateChanged(user)">修改状态</el-button>
-              <el-button type="danger" @click="deleteById(user.id)">删除用户</el-button>
+            <div class="op-col">
+              <span @click="userStateChanged(user)" class="op-text">修改状态</span>
+              <span @click="deleteById(user.id)" class="delete-text">删除</span>
+            </div>
           </td>
         </tr>
       </tbody>
     </table>
+
+     <div v-if="userlist===undefined||userlist.length===0">
+       <el-empty description="暂无内容"></el-empty>
+     </div>
 <!--     分页-->
      <el-pagination
          @size-change="handleSizeChange"
          @current-change="handleCurrentChange"
          :current-page="queryInfo.pagenum"
-         :page-sizes="[5, 10, 15]"
+         :page-sizes="[ 10, 15,30]"
          :page-size="queryInfo.pagesize"
          layout="total, sizes, prev, pager, next, jumper"
          :total="total"
@@ -64,37 +69,19 @@ export default {
         //当前页数
         pagenum: 1,
         // 每页显示多少数据
-        pagesize: 5
+        pagesize: 10,
+        query:''
       },
       userlist:[],
       total:0,
-      query:''
     }
   },
-  computed:{
-    sex(val){
-      if(val===1){
-        return "男"
-      }
-      else if(val===2){
-        return "女"
-      }
-      return "未知"
-    },
 
-    status(val){
-      if(val===1){
-        return '正常'
-      }else{
-        return '异常'
-      }
-    },
-  },
   methods:{
     getUserList(){
       this.$http.get("/admin/getUser",{params:this.queryInfo}).then(res=>{
-        this.userlist=res.data.data.userlist.list
-        this.total=res.data.data.userlist.total
+        this.userlist=res.data.data.list
+        this.total=res.data.data.total
       })
     },
     userStateChanged(user){
@@ -122,31 +109,15 @@ export default {
       this.queryInfo.pagenum = newNum
       this.getUserList()
     },
-    findOne(){
-      if(this.query===''||this.query.length===0){
-        this.$message.error("请输入搜索内容")
-      }
-      else{
-        let num=parseInt(this.query,10)
-        if(num>0){
-          this.$http.get("/admin/getOne/"+num).then(res=>{
-            this.userlist=res.data.data.userlist
-            this.total=this.userlist.length
-            this.queryInfo.pagenum=1
-            this.queryInfo.pagesize=5
-          })
-        }
-        else {
-          this.$message.error("搜索内容无效")
-        }
-        console.log(num)
-      }
-    },
     deleteById(id){
-      this.$http.delete('/admin/user/'+id).then(res=>{
-        this.$message.info(res.data.message)
-      })
-      this.userlist = this.userlist.filter(x=>x.id!==id)
+      this.$confirm('确定删除该用户吗')
+          .then(_ => {
+            this.$http.delete('/admin/user/'+id).then(res=>{
+              this.$message.info(res.data.message)
+              this.userlist = this.userlist.filter(x=>x.id!==id)
+            })
+          })
+          .catch(_ => {});
     },
     link(data){
       this.$router.push('/profile/'+data)
@@ -158,14 +129,28 @@ export default {
 }
 </script>
 <style scoped>
-.mytable{
-  min-height: 800px;
+.op-text{
+  color: #66b9fd;
+  cursor: pointer;
+}
+.op-text:hover{
+  color: blue;
+  text-decoration: underline;
+}
+.delete-text{
+  padding-left: 10px;
+  color: orangered;
+  cursor: pointer;
+}
+.delete-text:hover{
+  color: red;
+  text-decoration: underline;
 }
 .tableUint{
-  width: 150px;
+  text-align: center;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  max-width: 150px;
+  white-space: nowrap;
 }
-.unitAvatar{
-  width: 400px;
-}
-
 </style>
